Prevent infinite fetch call in Error&Loading case
The original reason that new fetch is called infinitely is that callback
of useIntersect hook is recreated repeatedly, and this make observer
observe and trigger callback infinitely.

Simple flow is like below.

Error occur -> useFetchHook is rerenderd (since currentResult is changed) ->
onIntersect of useIntersect is newly created and callback is rerendered
-> observer is rerendered (since callback is changed) -> observer
observe in useEffect -> callback is called -> Fetch is called and cycle
is repeated

We have several ways to fix this but have some problems like this.
1. Block fetching when error -> We cannot retry
2. Separate onIntersect callback to prevent rerendering -> We cannot
   check current fetch is loading or not

So, to prevent infinite call simply, we need to check isIntersecting state is
changed or not.

Issue: #40 Define state and action for each pagination state